requirements.txt
# requirements.txt

fastapi
uvicorn[standard]
jinja2
python-dotenv
requests
assemblyai
google-generativeai
websockets
shelve